:art: split manager into cache manager
